+++ packages/vendure-plugin-metrics/src/api/metrics.resolver.ts
@@ -1,21 +1,22 @@
 import { Args, Query, Resolver } from '@nestjs/graphql';
 import { Allow, Ctx, Permission, RequestContext } from '@vendure/core';
-import { MetricsService } from './metrics.service';
 import {
   AdvancedMetricSummary,
   AdvancedMetricSummaryInput,
+  AdvancedMetricType,
 } from '../ui/generated/graphql';
+import { MetricsService } from './metrics.service';
 
 @Resolver()
 export class MetricsResolver {
-  constructor(private service: MetricsService) {}
+  constructor(private readonly metricsService: MetricsService) {}
 
   @Query()
   @Allow(Permission.ReadOrder)
-  async advancedMetricSummary(
+  async advancedMetricSummaries(
     @Ctx() ctx: RequestContext,
     @Args('input') input: AdvancedMetricSummaryInput
   ): Promise<AdvancedMetricSummary[]> {
-    return this.service.getMetrics(ctx, input);
+    return this.metricsService.getMetrics(ctx, input);
   }
 }