+++ packages/vendure-plugin-metrics/src/ui/queries.graphql.ts
@@ -1,16 +1,16 @@
 import gql from 'graphql-tag';
 
 export const GET_METRICS = gql`
-  query advancedMetricSummary($input: AdvancedMetricSummaryInput!) {
-    advancedMetricSummary(input: $input) {
-      interval
+  query advancedMetricSummaries($input: AdvancedMetricSummaryInput) {
+    advancedMetricSummaries(input: $input) {
       code
       title
-      entries {
-        label
-        value
-      }
       type
+      labels
+      series {
+        name
+        values
+      }
     }
   }
 `;